requirements.txt
Flask
Flask-SQLAlchemy
Pillow
google-generativeai
python-dotenv

